+++ bindings/angular1/directives/input.js
@@ -56,20 +56,7 @@
 (function(){
   'use strict';
 
-  var ATTRS = [
-    'ngModel',
-    'ngChange',
-    'ngRequired',
-    'ngMinlength',
-    'ngMaxlength',
-    'ngPattern',
-    'ngTrim',
-    'ngValue',
-    'ngTrueValue',
-    'ngFalseValue'
-  ];
-
-  angular.module('onsen').directive('onsInput', function($compile) {
+  angular.module('onsen').directive('onsInput', function($parse) {
     return {
       restrict: 'E',
       replace: false,
@@ -77,29 +64,51 @@
 
       link: function(scope, element, attrs) {
         CustomElements.upgrade(element[0]);
+        let el = element[0];
 
-        var el = element[0];
-        var type = el.getAttribute('type');
+        const onInput = () => {
+          const set = $parse(attrs.ngModel).assign;
 
-        ATTRS.forEach(function(attr) {
-          var kebabCase = attr.replace(/[A-Z]/g, function(letter, pos) {
-            return (pos ? '-' : '') + letter.toLowerCase();
-          });
-
-          if (attrs.hasOwnProperty(attr)) {
-            el._input.setAttribute(kebabCase, attrs[attr]);
+          if (el._isTextInput) {
+            set(scope, el.value);
           }
-        });
+          else if (el.type === 'radio' && el.checked) {
+            set(scope, el.value);
+          }
+          else {
+            set(scope, el.checked);
+          }
 
-        $compile(el._input)(scope);
+          if (attrs.ngChange) {
+            scope.$eval(attrs.ngChange);
+          }
 
-        if (el._isTextInput && attrs.ngModel) {
-          scope.$watch(attrs.ngModel, function(value) {
-            el._updateLabelClass();
+          scope.$parent.$evalAsync();
+        };
+
+        if (attrs.ngModel) {
+          scope.$watch(attrs.ngModel, (value) => {
+            if (el._isTextInput) {
+              el.value = value;
+            }
+            else if (el.type === 'radio') {
+              el.checked = value === el.value;
+            }
+            else {
+              el.checked = value;
+            }
           });
+
+          el._isTextInput
+            ? element.on('input', onInput)
+            : element.on('change', onInput);
         }
 
-        scope.$on('$destroy', function() {
+        scope.$on('$destroy', () => {
+          el._isTextInput
+            ? element.off('input', onInput)
+            : element.off('change', onInput);
+
           scope = element = attrs = el = null;
         });
       }