+++ core/src/elements/ons-if.js
@@ -18,6 +18,7 @@ limitations under the License.
 import orientation from 'ons/orientation';
 import platform from 'ons/platform';
 import BaseElement from 'ons/base-element';
+import contentReady from 'ons/content-ready';
 
 /**
  * @element ons-if
@@ -65,11 +66,16 @@ class ConditionalElement extends BaseElement {
    */
 
   createdCallback() {
-    if (platform._renderPlatform !== null) {
-      this._platformUpdate();
-    } else if (!this._isAllowedPlatform()){
-      this.innerHTML = '';
-    }
+    contentReady(this, () => {
+      if (platform._renderPlatform !== null) {
+        this._platformUpdate();
+      } else if (!this._isAllowedPlatform()) {
+        while (this.childNodes[0]) {
+          this.childNodes[0].remove();
+        }
+        this._platformUpdate();
+      }
+    });
 
     this._onOrientationChange();
   }