+++ core/src/elements/ons-speed-dial.js
@@ -47,6 +47,20 @@ const scheme = {
 class SpeedDialElement extends BaseElement {
 
   /**
+   * @event open
+   * @description
+   *   [en]Fired when the menu items are shown.[/en]
+   *   [ja][/ja]
+   */
+
+  /**
+   * @event close
+   * @description
+   *   [en]Fired when the menu items are hidden.[/en]
+   *   [ja][/ja]
+   */
+
+  /**
    * @attribute modifier
    * @type {String}
    * @description
@@ -134,6 +148,7 @@ class SpeedDialElement extends BaseElement {
   }
 
   attachedCallback() {
+    this._updateDirection(this.hasAttribute('direction') ? this.getAttribute('direction') : 'up');
     this.addEventListener('click', this._boundOnClick, false);
   }
 
@@ -146,7 +161,7 @@ class SpeedDialElement extends BaseElement {
   }
 
   _onClick(e) {
-    if (!this.disabled) {
+    if (!this.disabled && this._shown) {
       this.toggleItems();
     }
   }
@@ -165,7 +180,10 @@ class SpeedDialElement extends BaseElement {
 
   _updateRipple() {
     const fab = util.findChild(this, 'ons-fab');
-    this.hasAttribute('ripple') ? fab.setAttribute('ripple', '') : fab.removeAttribute('ripple');
+
+    if (fab) {
+      this.hasAttribute('ripple') ? fab.setAttribute('ripple', '') : fab.removeAttribute('ripple');
+    }
   }
 
   _updateDirection(direction) {
@@ -292,6 +310,8 @@ class SpeedDialElement extends BaseElement {
       }
     }
     this._itemShown = true;
+
+    util.triggerElementEvent(this, 'open');
   }
 
   /**
@@ -312,6 +332,7 @@ class SpeedDialElement extends BaseElement {
       }
     }
     this._itemShown = false;
+    util.triggerElementEvent(this, 'close');
   }
 
   /**
@@ -360,7 +381,14 @@ class SpeedDialElement extends BaseElement {
     return this._shown && this.style.display !== 'none';
   }
 
-  isItemShown() {
+  /**
+   * @method isOpen
+   * @signature isOpen()
+   * @description
+   *   [en]Returns whether the menu is open or not.[/en]
+   *   [ja][/ja]
+   */
+  isOpen() {
     return this._itemShown;
   }
 
@@ -383,7 +411,7 @@ class SpeedDialElement extends BaseElement {
    *   [ja]Speed dialの子要素の表示非表示を切り替えます。[/ja]
    */
   toggleItems() {
-    if (this.isItemShown()) {
+    if (this.isOpen()) {
       this.hideItems();
     } else {
       this.showItems();